requirements.txt
h5py~=3.6.0
torch~=1.10.0
pytorch-lightning~=1.5.3
numpy~=1.21.4
matplotlib~=3.3.4
pyserial~=3.5
multiprocess~=0.70.12.2
tqdm~=4.62.2
ipython~=7.30.0
pyautogui~=0.9.53